refactor: reorganize types and utils (closes #42)
- Move validation functions (ValidateText, CheckIfUrlIsValid, IsImageURL,
  CheckIfValidTwitterURL, IsValidEthereumAddress, ReadPublicURL, etc.)
  from utils/ to types/validation.go where they belong
- Move validation errors and constants to types package
- Remove unused RoundUpDivideBig function from utils
- Make TypedErr unexported (typedErr) as internal helper
- Keep backward-compatible deprecated aliases in utils/errors.go
  so existing consumers don't break
- types/ no longer depends on utils/ for validation, breaking
  the potential circular dependency chain
