+++ types/operator_test.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/Layr-Labs/eigensdk-go/utils"
+	
 
 	"github.com/stretchr/testify/assert"
 )
@@ -44,7 +44,7 @@ func TestOperatorValidate(t *testing.T) {
 				MetadataUrl:               "",
 			},
 			wantErr:     true,
-			expectedErr: utils.WrapError(ErrInvalidMetadataUrl, utils.ErrEmptyUrl),
+			expectedErr: WrapError(ErrInvalidMetadataUrl, ErrEmptyUrl),
 		},
 		{
 			name: "failed operator validation - localhost metadata url",
@@ -54,7 +54,7 @@ func TestOperatorValidate(t *testing.T) {
 				MetadataUrl:               "http://localhost:8080/metadata.json",
 			},
 			wantErr:     true,
-			expectedErr: utils.WrapError(ErrInvalidMetadataUrl, utils.ErrUrlPointingToLocalServer),
+			expectedErr: WrapError(ErrInvalidMetadataUrl, ErrUrlPointingToLocalServer),
 		},
 		{
 			name: "failed operator validation - 127.0.0.1 metadata url",
@@ -64,7 +64,7 @@ func TestOperatorValidate(t *testing.T) {
 				MetadataUrl:               "http://127.0.0.1:8080/metadata.json",
 			},
 			wantErr:     true,
-			expectedErr: utils.WrapError(ErrInvalidMetadataUrl, utils.ErrUrlPointingToLocalServer),
+			expectedErr: WrapError(ErrInvalidMetadataUrl, ErrUrlPointingToLocalServer),
 		},
 		{
 			// TODO: avoid doing a real request in the test
@@ -75,7 +75,7 @@ func TestOperatorValidate(t *testing.T) {
 				MetadataUrl:               "https://github.com/layr-labs/-non-existent-repo/metadata.json",
 			},
 			wantErr: true,
-			expectedErr: utils.WrapError(
+			expectedErr: WrapError(
 				ErrReadingMetadataUrlResponse,
 				errors.New("error fetching url: 404 Not Found"),
 			),
