+++ utils/errors.go
@@ -1,59 +1,85 @@
+// Package utils provides utility functions.
+//
+// Deprecated: Validation functions and errors have been moved to the types package.
+// This file maintains backward compatibility. New code should use the types package directly.
 package utils
 
 import (
-	"errors"
-	"fmt"
-	"strings"
+	"github.com/Layr-Labs/eigensdk-go/types"
 )
 
-var (
-	ErrInvalidUrl          = errors.New("invalid url")
-	ErrInvalidGithubRawUrl = errors.New("invalid github raw url")
-	ErrInvalidText         = fmt.Errorf("invalid text format, doesn't conform to regex %s", TextRegex)
-	ErrTextTooLong         = func(limit int) error {
-		return fmt.Errorf("text should be less than %d characters", limit)
-	}
-	ErrEmptyText             = errors.New("text is empty")
-	ErrInvalidImageExtension = errors.New(
-		"invalid image extension. only " + strings.Join(ImageExtensions, ",") + " is supported",
-	)
-	ErrInvalidImageMimeType     = errors.New("invalid image mime-type. only png is supported")
-	ErrInvalidUrlLength         = errors.New("url length should be no larger than 1024 character")
-	ErrUrlPointingToLocalServer = errors.New("url should not point to local server")
-	ErrEmptyUrl                 = errors.New("url is empty")
-	ErrInvalidTwitterUrlRegex   = errors.New(
-		"invalid twitter url, it should be of the format https://twitter.com/<username> or https://x.com/<username>",
-	)
-	ErrResponseTooLarge = errors.New("response too large, allowed size is 1 MB")
-)
+// Deprecated: Use types.ErrInvalidUrl instead.
+var ErrInvalidUrl = types.ErrInvalidUrl
 
+// Deprecated: Use types.ErrInvalidGithubRawUrl instead.
+var ErrInvalidGithubRawUrl = types.ErrInvalidGithubRawUrl
+
+// Deprecated: Use types.ErrInvalidText instead.
+var ErrInvalidText = types.ErrInvalidText
+
+// Deprecated: Use types.ErrTextTooLong instead.
+var ErrTextTooLong = types.ErrTextTooLong
+
+// Deprecated: Use types.ErrEmptyText instead.
+var ErrEmptyText = types.ErrEmptyText
+
+// Deprecated: Use types.ErrInvalidImageExtension instead.
+var ErrInvalidImageExtension = types.ErrInvalidImageExtension
+
+// Deprecated: Use types.ErrInvalidImageMimeType instead.
+var ErrInvalidImageMimeType = types.ErrInvalidImageMimeType
+
+// Deprecated: Use types.ErrInvalidUrlLength instead.
+var ErrInvalidUrlLength = types.ErrInvalidUrlLength
+
+// Deprecated: Use types.ErrUrlPointingToLocalServer instead.
+var ErrUrlPointingToLocalServer = types.ErrUrlPointingToLocalServer
+
+// Deprecated: Use types.ErrEmptyUrl instead.
+var ErrEmptyUrl = types.ErrEmptyUrl
+
+// Deprecated: Use types.ErrInvalidTwitterUrlRegex instead.
+var ErrInvalidTwitterUrlRegex = types.ErrInvalidTwitterUrlRegex
+
+// Deprecated: Use types.ErrResponseTooLarge instead.
+var ErrResponseTooLarge = types.ErrResponseTooLarge
+
+// Deprecated: Use types.TextCharsLimit instead.
+var TextCharsLimit = types.TextCharsLimit
+
+// Deprecated: Use types.ImageExtensions instead.
+var ImageExtensions = types.ImageExtensions
+
+// Deprecated: Use types.WrapError instead.
+var WrapError = types.WrapError
+
+// Deprecated: Use types.IsValidEthereumAddress instead.
+var IsValidEthereumAddress = types.IsValidEthereumAddress
+
+// Deprecated: Use types.ReadPublicURL instead.
+var ReadPublicURL = types.ReadPublicURL
+
+// Deprecated: Use types.CheckIfValidTwitterURL instead.
+var CheckIfValidTwitterURL = types.CheckIfValidTwitterURL
+
+// Deprecated: Use types.CheckBasicURLValidation instead.
+var CheckBasicURLValidation = types.CheckBasicURLValidation
+
+// Deprecated: Use types.CheckIfUrlIsValid instead.
+var CheckIfUrlIsValid = types.CheckIfUrlIsValid
+
+// Deprecated: Use types.IsImageURL instead.
+var IsImageURL = types.IsImageURL
+
+// Deprecated: Use types.ValidateText instead.
+var ValidateText = types.ValidateText
+
+// Deprecated: Use types.ValidateRawGithubUrl instead.
+var ValidateRawGithubUrl = types.ValidateRawGithubUrl
+
+// TypedErr is deprecated. Use types.WrapError instead.
+//
+// Deprecated: This function is deprecated and will be removed in the future.
 func TypedErr(e interface{}) error {
-	switch t := e.(type) {
-	case error:
-		return t
-	case string:
-		return errors.New(t)
-	default:
-		return nil
-	}
-}
-
-func WrapError(mainErr interface{}, subErr interface{}) error {
-	var main, sub error
-	main = TypedErr(mainErr)
-	sub = TypedErr(subErr)
-	// Some times the wrap will wrap a nil error
-	if main == nil && sub == nil {
-		return nil
-	}
-
-	if main == nil && sub != nil {
-		return sub
-	}
-
-	if main != nil && sub == nil {
-		return main
-	}
-
-	return fmt.Errorf("%w: %w", main, sub)
+	return types.WrapError(e, nil)
 }
